refactor and commenting
