+++ src/components/substance/agSubstanceRelationshipTable.vue
@@ -25,7 +25,7 @@ import {
 } from "@/ag-grid-components/custom-renderers";
 
 export default {
-  name: "SynonymTable",
+  name: "agSubstanceRelationshipTable",
   components: {
     AgGridVue
   },
